[chore] : 디렉토리 변경 data-dto-remote
