+++ app/src/main/java/org/sopt/dosopttemplate/data/dto/remote/request/RequestLoginDto.kt
@@ -1,14 +1,12 @@
-package org.sopt.dosopttemplate.network.login
+package org.sopt.dosopttemplate.data.dto.remote.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class RequestSignUpDto(
+data class RequestLoginDto(
     @SerialName("username")
     val username: String,
     @SerialName("password")
     val password: String,
-    @SerialName("nickname")
-    val nickname: String,
 )
