+++ app/src/main/java/org/sopt/dosopttemplate/data/dto/remote/respose/ResponseLoginDto.kt
@@ -1,0 +1,14 @@
+package org.sopt.dosopttemplate.data.dto.remote.respose
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ResponseLoginDto(
+    @SerialName("id")
+    val id: Int,
+    @SerialName("username")
+    val username: String,
+    @SerialName("nickname")
+    val nickname: String? = null,
+)
