--- app/src/main/java/org/sopt/dosopttemplate/network/login/RequestLoginDto.kt
@@ -1,12 +1,0 @@
-package org.sopt.dosopttemplate.network.login
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class RequestLoginDto(
-    @SerialName("username")
-    val username: String,
-    @SerialName("password")
-    val password: String,
-)
